# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_gateway_request.CreateGatewayRequest:
    properties:
      client_tls:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_domain_gateway.ClientTLSConfig'
      domain:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      session_config:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_domain_gateway.SessionConfig'
      slug:
        type: string
      telemetry:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_domain_telemetry.Telemetry'
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_gateway_response.GatewayHosts:
    properties:
      mcp:
        type: string
      proxy:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_gateway_response.GatewayResponse:
    properties:
      client_tls:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_domain_gateway.ClientTLSConfig'
      created_at:
        type: string
      domain:
        type: string
      hosts:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_gateway_response.GatewayHosts'
      id:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      session_config:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_domain_gateway.SessionConfig'
      slug:
        type: string
      status:
        type: string
      telemetry:
        $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_domain_telemetry.Telemetry'
      updated_at:
        type: string
      version:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_gateway_response.ListGatewayResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_gateway_response.GatewayResponse'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_policy_request.CreatePolicyRequest:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      mode:
        type: string
      name:
        type: string
      parallel:
        type: boolean
      priority:
        type: integer
      settings:
        additionalProperties: {}
        type: object
      slug:
        type: string
      stages:
        items:
          type: string
        type: array
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_policy_response.ListPolicyResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_policy_response.PolicyResponse'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  github_com_NeuralTrust_TrustGate_pkg_api_handler_http_policy_response.PolicyResponse:
    properties:
      consumer_ids:
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      gateway_id:
        type: string
      global:
        type: boolean
      id:
        type: string
      mode:
        type: string
      name:
        type: string
      parallel:
        type: boolean
      priority:
        type: integer
      settings:
        additionalProperties: {}
        type: object
      slug:
        type: string
      stages:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_domain_gateway.ClientTLSConfig:
    additionalProperties:
      items:
        type: integer
      type: array
    type: object
  github_com_NeuralTrust_TrustGate_pkg_domain_gateway.SessionConfig:
    properties:
      body_param_name:
        type: string
      enabled:
        type: boolean
      header_name:
        type: string
    type: object
...
  github_com_NeuralTrust_TrustGate_pkg_domain_telemetry.ExporterConfig:
    properties:
      name:
        type: string
      settings:
        additionalProperties: true
        type: object
    type: object
  github_com_NeuralTrust_TrustGate_pkg_domain_telemetry.Telemetry:
    properties:
      enable_plugin_traces:
        type: boolean
      enable_request_traces:
        type: boolean
      exporters:
        items:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_domain_telemetry.ExporterConfig'
        type: array
      extra_params:
        additionalProperties:
          type: string
        type: object
      header_mapping:
        additionalProperties:
          type: string
        type: object
    type: object
...
info:
  contact:
    email: support@neuraltrust.ai
    name: NeuralTrust
    url: https://neuraltrust.ai/contact
  description: Administrative API for managing gateways and their registries, policies,
    consumers, roles and auth credentials.
  title: TrustGate Admin API
  version: "1.0"
paths:
...
  /v1/gateways:
    get:
      description: Returns a paginated list of gateways.
      parameters:
      - description: Filter by name (substring match)
        in: query
        name: name
        type: string
      - description: Page number (1-based)
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_gateway_response.ListGatewayResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: List gateways
      tags:
      - gateways
    post:
      consumes:
      - application/json
      description: Creates a new gateway.
      parameters:
      - description: Gateway to create
        in: body
        name: gateway
        required: true
        schema:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_gateway_request.CreateGatewayRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_gateway_response.GatewayResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create a gateway
      tags:
      - gateways
...
  /v1/gateways/{gateway_id}/policies:
    get:
      description: Returns a paginated list of policies in a gateway.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Filter by name (substring match)
        in: query
        name: name
        type: string
      - description: Page number (1-based)
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_policy_response.ListPolicyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: List policies
      tags:
      - policies
    post:
      consumes:
      - application/json
      description: Creates a new policy in a gateway.
      parameters:
      - description: Gateway id
        format: uuid
        in: path
        name: gateway_id
        required: true
        type: string
      - description: Policy to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_policy_request.CreatePolicyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_policy_response.PolicyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_NeuralTrust_TrustGate_pkg_api_handler_http_helpers.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create a policy
      tags:
      - policies
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"